backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
openai
pinecone-client
stripe
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-decouple
aiofiles
PyJWT
pinecone